+++ tests/testthat/test-collapse.R
@@ -21,3 +21,17 @@ test_that("can collapse missing levels", {
 
   expect_equal(f2, factor(c("x", "y")))
 })
+
+test_that("can collapse un-named levels to Other", {
+  f1 <- factor(letters[1:3])
+  f2 <- fct_collapse(f1, xy = c("a", "b"), group_other = TRUE)
+
+  expect_equal(f2, factor(c("xy", "xy", "Other"), levels = c("xy", "Other")))
+})
+
+test_that("does not automatically collapse unnamed levels to Other", {
+  f1 <- factor(letters[1:3])
+  f2 <- fct_collapse(f1, xy = c("a", "b"))
+
+  expect_equal(f2, factor(c("xy", "xy", "c"), levels = c("xy", "c")))
+})